Create readme
The major part of a Drone is a microcomputer that is programmable and electric motors for turning the blades.Thus a drone architecture is simple  which makes it an electrical device. However the control system for it can be very expensive. Web based control for the Drone is very cheap but may not be as effective as hardware FPGA based. The type of control system of a drone depends on the use of a drone.for use in agricultural systems a best effort service may suffice.